pyproject.toml
[project]
name = "konsilisyum"
version = "0.1.0"
description = "Terminal tabanli yasayan fikir meclisi"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
konsilisyum = "konsilisyum.main:main"

[tool.setuptools.packages.find]
include = ["konsilisyum*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
